pyproject.toml
[project]
name = "othello-bot"
version = "0.1.0"
description = "Machine learning for playing Othello"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
]
